refactor: changed cryptoStore to assetStore
