+++ Expensier.WPF/HostBuilders/AddStoreHostBuilderExtension.cs
@@ -11,7 +11,7 @@ using System.Text;
 using System.Threading.Tasks;
 using Expensier.WPF.ViewModels;
 using Expensier.WPF.State.Subscriptions;
-using Expensier.WPF.State.Crypto;
+using Expensier.WPF.State.Assets;
 using Expensier.WPF.State.Portfolio;
 
 namespace Expensier.WPF.HostBuilders
@@ -28,7 +28,7 @@ namespace Expensier.WPF.HostBuilders
                 services.AddSingleton<AccountStore, AccountStore>();
                 services.AddSingleton<TransactionStore, TransactionStore>();
                 services.AddSingleton<SubscriptionStore, SubscriptionStore>();
-                services.AddSingleton<CryptoStore, CryptoStore>();
+                services.AddSingleton<AssetStore, AssetStore>();
                 services.AddSingleton<PortfolioStore, PortfolioStore>();
             });
