+++ Expensier.WPF/ViewModels/Cryptos/CryptoWatchlistViewModel.cs
@@ -1,5 +1,5 @@
 ﻿using Expensier.Domain.Services;
-using Expensier.WPF.State.Crypto;
+using Expensier.WPF.State.Assets;
 using Microsoft.IdentityModel.Tokens;
 using System;
 using System.Collections.Generic;
@@ -12,7 +12,7 @@ namespace Expensier.WPF.ViewModels.Cryptos
 {
     public class CryptoWatchlistViewModel : CryptoWatchlistBaseViewModel
     {
-        public CryptoWatchlistViewModel(CryptoStore cryptoStore, IAssetService cryptoService)
+        public CryptoWatchlistViewModel(AssetStore cryptoStore, IAssetService cryptoService)
             : base(cryptoStore, cryptoService, cryptos => cryptos.OrderByDescending(c => c.Asset.PercentageChange)) { }
     }
 }
